refactor: unsigned int -> size_t

+++ io.cpp
@@ -4,14 +4,14 @@
  * поиск элемента `searched` начиная от текущей позиции итератора
  * возвращает кол-во элементов между текущим элементом и искомым
  */
-unsigned int findElementNext(std::vector<SparseDoubleLinkedMatrixElement *>::iterator current,
+size_t findElementNext(std::vector<SparseDoubleLinkedMatrixElement *>::iterator current,
                              const SparseDoubleLinkedMatrixElement *searched) {
     int i = 0;
     for (; *current != searched; ++current, ++i) {}
     return i;
 }
 
-SparseDoubleLinkedMatrixElement *createIfNotExists(std::unordered_map<unsigned int, SparseDoubleLinkedMatrixElement *>& table, unsigned int id,
+SparseDoubleLinkedMatrixElement *createIfNotExists(std::unordered_map<size_t, SparseDoubleLinkedMatrixElement *>& table, size_t id,
                                                    const double value = NAN) {
     if(id == 0)
         return nullptr;
@@ -24,7 +24,7 @@ SparseDoubleLinkedMatrixElement *createIfNotExists(std::unordered_map<unsigned i
     return table.at(id);
 }
 
-void loadPointers(const std::vector<std::string>& line, std::unordered_map<unsigned int, SparseDoubleLinkedMatrixElement*>& elements, std::vector<SparseDoubleLinkedMatrixElement *>& pointers) {
+void loadPointers(const std::vector<std::string>& line, std::unordered_map<size_t, SparseDoubleLinkedMatrixElement*>& elements, std::vector<SparseDoubleLinkedMatrixElement *>& pointers) {
     for (const auto& number: line) {
         int id = std::stoi(number);
         if(id != 0) {
@@ -36,7 +36,7 @@ void loadPointers(const std::vector<std::string>& line, std::unordered_map<unsig
     }
 }
 
-void loadElement(const std::vector<std::string>& line, std::unordered_map<unsigned int, SparseDoubleLinkedMatrixElement*>& elements, int id) {
+void loadElement(const std::vector<std::string>& line, std::unordered_map<size_t, SparseDoubleLinkedMatrixElement*>& elements, int id) {
 
     const double value = std::stod(line[0]);
     const int nextLineId = std::stoi(line[1]), nextColumnId = std::stoi(line[2]);
@@ -56,7 +56,7 @@ SparseDoubleLinkedMatrix loadFromFile(const std::string &path) {
     const auto lines = countLines(path);
     if(lines < 3)
         return matrix;
-    std::unordered_map<unsigned int, SparseDoubleLinkedMatrixElement*> elements(lines-2);
+    std::unordered_map<size_t, SparseDoubleLinkedMatrixElement*> elements(lines-2);
 
     std::getline(input,lineString);
     std::vector<std::string> line = split(lineString, DLSMDelimiter);
@@ -77,14 +77,14 @@ SparseDoubleLinkedMatrix loadFromFile(const std::string &path) {
 
 void saveToFile(const std::string &path, const SparseDoubleLinkedMatrix& matrix) {
     std::vector<SparseDoubleLinkedMatrixElement*> elements(matrix.linePointer.size()*matrix.columnPointer.size() / 10);
-    std::vector<unsigned int> lineIds(matrix.linePointer.size(), 0), columnIds(matrix.columnPointer.size(), 0);
-    unsigned int k = 1, i = 0;
+    std::vector<size_t> lineIds(matrix.linePointer.size(), 0), columnIds(matrix.columnPointer.size(), 0);
+    size_t k = 1, i = 0;
 
     std::vector<SparseDoubleLinkedMatrixElement*> columnPointer(matrix.columnPointer.size());
     std::ranges::copy(matrix.columnPointer, columnPointer.begin());
     for(auto lineHead = matrix.linePointer.begin(); lineHead != matrix.linePointer.end(); ++i, ++lineHead){
         auto lineTail = *lineHead;
-        unsigned int j = 0;
+        size_t j = 0;
         for (auto columnHead = columnPointer.begin(); columnHead != columnPointer.end(); ++j, ++columnHead) {
             if(lineTail && lineTail == *columnHead) {
                 elements[k-1] = (*columnHead);
