+++ util.h
@@ -4,7 +4,7 @@
 #include "main.h"
 
 std::vector<std::string> split(const std::string &s, char delim);
-std::string join(const std::vector<unsigned int> &lst, char delim);
-unsigned int countLines(const std::string& path);
+std::string join(const std::vector<size_t> &lst, char delim);
+size_t countLines(const std::string& path);
 
 #endif //UTIL_H
